Refactored myPet into its own .py file and updated tests
